Migrate applied rules page
