+++ modules/features/filters/src/main/java/au/com/shiftyjelly/pocketcasts/playlists/rules/AppliedRulesPage.kt
@@ -1,7 +1,5 @@
-package au.com.shiftyjelly.pocketcasts.playlists.create
+package au.com.shiftyjelly.pocketcasts.playlists.rules
 
-import androidx.annotation.DrawableRes
-import androidx.annotation.StringRes
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.core.animateFloatAsState
 import androidx.compose.foundation.Image
@@ -14,6 +12,7 @@ import androidx.compose.foundation.layout.IntrinsicSize
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.WindowInsets
 import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -53,6 +52,8 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import au.com.shiftyjelly.pocketcasts.compose.AppThemeWithBackground
 import au.com.shiftyjelly.pocketcasts.compose.Devices
+import au.com.shiftyjelly.pocketcasts.compose.bars.NavigationButton
+import au.com.shiftyjelly.pocketcasts.compose.bars.ThemedTopAppBar
 import au.com.shiftyjelly.pocketcasts.compose.buttons.RowButton
 import au.com.shiftyjelly.pocketcasts.compose.components.EpisodeImage
 import au.com.shiftyjelly.pocketcasts.compose.components.FadedLazyColumn
@@ -74,7 +75,6 @@ import au.com.shiftyjelly.pocketcasts.models.type.SmartRules.EpisodeDurationRule
 import au.com.shiftyjelly.pocketcasts.models.type.SmartRules.MediaTypeRule
 import au.com.shiftyjelly.pocketcasts.models.type.SmartRules.PodcastsRule
 import au.com.shiftyjelly.pocketcasts.models.type.SmartRules.ReleaseDateRule
-import au.com.shiftyjelly.pocketcasts.playlists.create.CreatePlaylistViewModel.AppliedRules
 import au.com.shiftyjelly.pocketcasts.repositories.extensions.getSummaryText
 import au.com.shiftyjelly.pocketcasts.repositories.images.PocketCastsImageRequestFactory.PlaceholderType
 import au.com.shiftyjelly.pocketcasts.ui.extensions.getThemeColor
@@ -84,48 +84,14 @@ import au.com.shiftyjelly.pocketcasts.images.R as IR
 import au.com.shiftyjelly.pocketcasts.localization.R as LR
 import au.com.shiftyjelly.pocketcasts.ui.R as UR
 
-enum class RuleType(
-    @DrawableRes val iconId: Int,
-    @StringRes val titleId: Int,
-) {
-    Podcasts(
-        iconId = IR.drawable.ic_rule_podcasts,
-        titleId = LR.string.podcasts,
-    ),
-    EpisodeStatus(
-        iconId = IR.drawable.ic_rule_episode_status,
-        titleId = LR.string.filters_chip_episode_status,
-    ),
-    ReleaseDate(
-        iconId = IR.drawable.ic_rule_release_date,
-        titleId = LR.string.filters_release_date,
-    ),
-    EpisodeDuration(
-        iconId = IR.drawable.ic_rule_duration,
-        titleId = LR.string.filters_duration,
-    ),
-    DownloadStatus(
-        iconId = IR.drawable.ic_rule_download_status,
-        titleId = LR.string.filters_chip_download_status,
-    ),
-    MediaType(
-        iconId = IR.drawable.ic_rule_media_type,
-        titleId = LR.string.filters_chip_media_type,
-    ),
-    Starred(
-        iconId = IR.drawable.ic_rule_starred,
-        titleId = LR.string.filters_chip_starred,
-    ),
-}
-
 @Composable
-fun SmartPlaylistPreviewPage(
+fun AppliedRulesPage(
     playlistTitle: String,
     appliedRules: AppliedRules,
     availableEpisodes: List<PodcastEpisode>,
     useEpisodeArtwork: Boolean,
     areOtherOptionsExpanded: Boolean,
-    onCreateSmartPlaylist: () -> Unit,
+    onCreatePlaylist: () -> Unit,
     onClickRule: (RuleType) -> Unit,
     toggleOtherOptions: () -> Unit,
     onClickClose: () -> Unit,
@@ -145,6 +111,13 @@ fun SmartPlaylistPreviewPage(
                 tint = MaterialTheme.theme.colors.primaryIcon03,
             )
         }
+        ThemedTopAppBar(
+            navigationButton = NavigationButton.Close,
+            style = ThemedTopAppBar.Style.Immersive,
+            iconColor = MaterialTheme.theme.colors.primaryIcon03,
+            windowInsets = WindowInsets(0),
+            onNavigationClick = onClickClose,
+        )
         Column(
             modifier = Modifier.weight(1f),
         ) {
@@ -233,7 +206,7 @@ fun SmartPlaylistPreviewPage(
             RowButton(
                 text = stringResource(LR.string.create_smart_playlist),
                 enabled = appliedRules.isAnyRuleApplied,
-                onClick = onCreateSmartPlaylist,
+                onClick = onCreatePlaylist,
                 includePadding = false,
                 modifier = Modifier
                     .padding(horizontal = 16.dp)
@@ -626,18 +599,18 @@ private fun AppliedRules.description(ruleType: RuleType) = when (ruleType) {
 
 @Preview(device = Devices.PORTRAIT_REGULAR)
 @Composable
-private fun SmartPlaylistsPreviewNoRulesPreview(
+private fun AppliedRulesPageNoRulesPreview(
     @PreviewParameter(ThemePreviewParameterProvider::class) themeType: ThemeType,
 ) {
     var expanded by remember { mutableStateOf(false) }
     AppThemeWithBackground(themeType) {
-        SmartPlaylistPreviewPage(
+        AppliedRulesPage(
             playlistTitle = "Comedy",
             appliedRules = AppliedRules.Empty,
             availableEpisodes = emptyList(),
             useEpisodeArtwork = false,
             areOtherOptionsExpanded = expanded,
-            onCreateSmartPlaylist = {},
+            onCreatePlaylist = {},
             onClickRule = {},
             onClickClose = {},
             toggleOtherOptions = { expanded = !expanded },
@@ -648,12 +621,12 @@ private fun SmartPlaylistsPreviewNoRulesPreview(
 
 @Preview(device = Devices.PORTRAIT_REGULAR)
 @Composable
-private fun SmartPlaylistsPreviewEpisodessPreview(
+private fun AppliedRulesPageEpisodessPreview(
     @PreviewParameter(ThemePreviewParameterProvider::class) themeType: ThemeType,
 ) {
     var expanded by remember { mutableStateOf(false) }
     AppThemeWithBackground(themeType) {
-        SmartPlaylistPreviewPage(
+        AppliedRulesPage(
             playlistTitle = "Comedy",
             appliedRules = AppliedRules.Empty.copy(
                 podcasts = PodcastsRule.Any,
@@ -668,7 +641,7 @@ private fun SmartPlaylistsPreviewEpisodessPreview(
             },
             useEpisodeArtwork = false,
             areOtherOptionsExpanded = expanded,
-            onCreateSmartPlaylist = {},
+            onCreatePlaylist = {},
             onClickRule = {},
             onClickClose = {},
             toggleOtherOptions = { expanded = !expanded },
@@ -679,12 +652,12 @@ private fun SmartPlaylistsPreviewEpisodessPreview(
 
 @Preview(device = Devices.PORTRAIT_REGULAR)
 @Composable
-private fun SmartPlaylistsPreviewNoEpisodesPreview(
+private fun AppliedRulesPageNoEpisodesPreview(
     @PreviewParameter(ThemePreviewParameterProvider::class) themeType: ThemeType,
 ) {
     var expanded by remember { mutableStateOf(false) }
     AppThemeWithBackground(themeType) {
-        SmartPlaylistPreviewPage(
+        AppliedRulesPage(
             playlistTitle = "Comedy",
             appliedRules = AppliedRules.Empty.copy(
                 podcasts = PodcastsRule.Any,
@@ -692,7 +665,7 @@ private fun SmartPlaylistsPreviewNoEpisodesPreview(
             availableEpisodes = emptyList(),
             useEpisodeArtwork = false,
             areOtherOptionsExpanded = expanded,
-            onCreateSmartPlaylist = {},
+            onCreatePlaylist = {},
             onClickRule = {},
             onClickClose = {},
             toggleOtherOptions = { expanded = !expanded },
