+++ modules/features/filters/src/main/java/au/com/shiftyjelly/pocketcasts/playlists/rules/RuleRadioRow.kt
@@ -10,7 +10,6 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.semantics.Role
 import androidx.compose.ui.unit.dp
 import au.com.shiftyjelly.pocketcasts.compose.components.TextH30
-import au.com.shiftyjelly.pocketcasts.playlists.create.RuleType
 
 @Composable
 fun RuleRadioRow(
